requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
evaluate>=0.4.0
sentencepiece>=0.1.99  # For XLM-R tokenizer

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Logging and experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# API clients (for LLM Teacher)
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: for local LLM
# torch-serve
# vllm

